+++ compiler-core/src/error/tests.rs
@@ -65,7 +65,20 @@ fn io_link_file_error_windows() {
         kind: FileKind::File,
         action: FileIoAction::Link("/dest".into()),
         path: "/src".into(),
-        err: Some("Critical error!".to_owned()),
+        err: FileIoFailure::Other("Critical error!".to_owned()),
+    }
+    .pretty_string();
+    assert_snapshot!(error);
+}
+
+#[cfg(windows)]
+#[test]
+fn io_link_file_error_not_found_windows() {
+    let error = Error::FileIo {
+        kind: FileKind::File,
+        action: FileIoAction::Link("/dest".into()),
+        path: "/src".into(),
+        err: FileIoFailure::NotFound,
     }
     .pretty_string();
     assert_snapshot!(error);
@@ -78,7 +91,20 @@ fn io_link_file_error_not_windows() {
         kind: FileKind::File,
         action: FileIoAction::Link("/dest".into()),
         path: "/src".into(),
-        err: Some("Critical error!".to_owned()),
+        err: FileIoFailure::Other("Critical error!".to_owned()),
+    }
+    .pretty_string();
+    assert_snapshot!(error);
+}
+
+#[cfg(not(windows))]
+#[test]
+fn io_link_file_error_not_found_not_windows() {
+    let error = Error::FileIo {
+        kind: FileKind::File,
+        action: FileIoAction::Link("/dest".into()),
+        path: "/src".into(),
+        err: FileIoFailure::NotFound,
     }
     .pretty_string();
     assert_snapshot!(error);
@@ -90,7 +116,19 @@ fn io_copy_directory_error() {
         kind: FileKind::Directory,
         action: FileIoAction::Copy(Some("/dest".into())),
         path: "/src".into(),
-        err: Some("Critical error!".to_owned()),
+        err: FileIoFailure::Other("Critical error!".to_owned()),
+    }
+    .pretty_string();
+    assert_snapshot!(error);
+}
+
+#[test]
+fn io_copy_directory_not_found_error() {
+    let error = Error::FileIo {
+        kind: FileKind::Directory,
+        action: FileIoAction::Copy(Some("/dest".into())),
+        path: "/src".into(),
+        err: FileIoFailure::NotFound,
     }
     .pretty_string();
     assert_snapshot!(error);
@@ -102,7 +140,19 @@ fn io_delete_file_error() {
         kind: FileKind::File,
         action: FileIoAction::Delete,
         path: "/file".into(),
-        err: Some("Critical error!".to_owned()),
+        err: FileIoFailure::Other("Critical error!".to_owned()),
+    }
+    .pretty_string();
+    assert_snapshot!(error);
+}
+
+#[test]
+fn io_delete_file_not_found_error() {
+    let error = Error::FileIo {
+        kind: FileKind::File,
+        action: FileIoAction::Delete,
+        path: "/file".into(),
+        err: FileIoFailure::NotFound,
     }
     .pretty_string();
     assert_snapshot!(error);
@@ -114,7 +164,19 @@ fn io_open_file_error() {
         kind: FileKind::File,
         action: FileIoAction::Open,
         path: "/file".into(),
-        err: Some("Critical error!".to_owned()),
+        err: FileIoFailure::Other("Critical error!".to_owned()),
+    }
+    .pretty_string();
+    assert_snapshot!(error);
+}
+
+#[test]
+fn io_open_file_not_found_error() {
+    let error = Error::FileIo {
+        kind: FileKind::File,
+        action: FileIoAction::Open,
+        path: "/file".into(),
+        err: FileIoFailure::NotFound,
     }
     .pretty_string();
     assert_snapshot!(error);
@@ -126,7 +188,19 @@ fn io_parse_file_error() {
         kind: FileKind::File,
         action: FileIoAction::Parse,
         path: "/file".into(),
-        err: Some("Critical error!".to_owned()),
+        err: FileIoFailure::Other("Critical error!".to_owned()),
+    }
+    .pretty_string();
+    assert_snapshot!(error);
+}
+
+#[test]
+fn io_parse_file_not_found_error() {
+    let error = Error::FileIo {
+        kind: FileKind::File,
+        action: FileIoAction::Parse,
+        path: "/file".into(),
+        err: FileIoFailure::NotFound,
     }
     .pretty_string();
     assert_snapshot!(error);
@@ -138,7 +212,19 @@ fn io_read_file_error() {
         kind: FileKind::File,
         action: FileIoAction::Read,
         path: "/file".into(),
-        err: Some("Critical error!".to_owned()),
+        err: FileIoFailure::Other("Critical error!".to_owned()),
+    }
+    .pretty_string();
+    assert_snapshot!(error);
+}
+
+#[test]
+fn io_read_file_not_found_error() {
+    let error = Error::FileIo {
+        kind: FileKind::File,
+        action: FileIoAction::Read,
+        path: "/file".into(),
+        err: FileIoFailure::NotFound,
     }
     .pretty_string();
     assert_snapshot!(error);
@@ -150,7 +236,19 @@ fn io_create_directory_error() {
         kind: FileKind::Directory,
         action: FileIoAction::Create,
         path: "/dir".into(),
-        err: Some("Critical error!".to_owned()),
+        err: FileIoFailure::Other("Critical error!".to_owned()),
+    }
+    .pretty_string();
+    assert_snapshot!(error);
+}
+
+#[test]
+fn io_create_directory_not_found_error() {
+    let error = Error::FileIo {
+        kind: FileKind::Directory,
+        action: FileIoAction::Create,
+        path: "/dir".into(),
+        err: FileIoFailure::NotFound,
     }
     .pretty_string();
     assert_snapshot!(error);
@@ -162,7 +260,19 @@ fn io_write_to_file_error() {
         kind: FileKind::File,
         action: FileIoAction::WriteTo,
         path: "/file".into(),
-        err: Some("Critical error!".to_owned()),
+        err: FileIoFailure::Other("Critical error!".to_owned()),
+    }
+    .pretty_string();
+    assert_snapshot!(error);
+}
+
+#[test]
+fn io_write_to_file_not_found_error() {
+    let error = Error::FileIo {
+        kind: FileKind::File,
+        action: FileIoAction::WriteTo,
+        path: "/file".into(),
+        err: FileIoFailure::NotFound,
     }
     .pretty_string();
     assert_snapshot!(error);
@@ -174,7 +284,19 @@ fn io_find_parent_of_directory_error() {
         kind: FileKind::Directory,
         action: FileIoAction::FindParent,
         path: "/dir".into(),
-        err: Some("Critical error!".to_owned()),
+        err: FileIoFailure::Other("Critical error!".to_owned()),
+    }
+    .pretty_string();
+    assert_snapshot!(error);
+}
+
+#[test]
+fn io_find_parent_of_directory_not_found_error() {
+    let error = Error::FileIo {
+        kind: FileKind::Directory,
+        action: FileIoAction::FindParent,
+        path: "/dir".into(),
+        err: FileIoFailure::NotFound,
     }
     .pretty_string();
     assert_snapshot!(error);
@@ -186,7 +308,19 @@ fn io_canonicalise_file_error() {
         kind: FileKind::File,
         action: FileIoAction::Canonicalise,
         path: "/file".into(),
-        err: Some("Critical error!".to_owned()),
+        err: FileIoFailure::Other("Critical error!".to_owned()),
+    }
+    .pretty_string();
+    assert_snapshot!(error);
+}
+
+#[test]
+fn io_canonicalise_file_not_found_error() {
+    let error = Error::FileIo {
+        kind: FileKind::File,
+        action: FileIoAction::Canonicalise,
+        path: "/file".into(),
+        err: FileIoFailure::NotFound,
     }
     .pretty_string();
     assert_snapshot!(error);
@@ -198,7 +332,19 @@ fn io_update_file_permissions_error() {
         kind: FileKind::File,
         action: FileIoAction::UpdatePermissions,
         path: "/file".into(),
-        err: Some("Critical error!".to_owned()),
+        err: FileIoFailure::Other("Critical error!".to_owned()),
+    }
+    .pretty_string();
+    assert_snapshot!(error);
+}
+
+#[test]
+fn io_update_file_permissions_not_found_error() {
+    let error = Error::FileIo {
+        kind: FileKind::File,
+        action: FileIoAction::UpdatePermissions,
+        path: "/file".into(),
+        err: FileIoFailure::NotFound,
     }
     .pretty_string();
     assert_snapshot!(error);
@@ -210,7 +356,19 @@ fn io_read_metadata_of_file_error() {
         kind: FileKind::File,
         action: FileIoAction::ReadMetadata,
         path: "/file".into(),
-        err: Some("Critical error!".to_owned()),
+        err: FileIoFailure::Other("Critical error!".to_owned()),
+    }
+    .pretty_string();
+    assert_snapshot!(error);
+}
+
+#[test]
+fn io_read_metadata_of_file_not_found_error() {
+    let error = Error::FileIo {
+        kind: FileKind::File,
+        action: FileIoAction::ReadMetadata,
+        path: "/file".into(),
+        err: FileIoFailure::NotFound,
     }
     .pretty_string();
     assert_snapshot!(error);